+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improve fallback functionality for Edge driver that in case of failure of downloading auto/latest driver tries to download driver of the closest version.
+
 ## [2.12.0] - 2024-07-30
 
 ### Added